requirements.txt
aiodns==2.0.0
aiohttp==3.8.1
aiohttp-apispec==2.2.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.7.7
anyio==3.2.1
apipkg==1.5
apispec==3.3.1
appdirs==1.4.4
appnope==0.1.0
arch==5.1.0
argon2-cffi==20.1.0
aries-cloudagent==0.5.3
asgiref==3.5.0
asn1crypto==1.3.0
astroid==2.4.1
asttokens==2.0.5
async-timeout==4.0.1
asyncio==3.4.3
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
backtrader==1.9.76.123
base58==2.1.1
binance-futures-connector==1.10.0
bitarray==1.2.2
black==21.11b1
bleach==3.1.0
brownie-token-tester==0.2.0
cached-property==1.5.2
cachetools==5.0.0
cctx==1.0.0
ccxt==1.53.14
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
conda==4.3.16
constantly==15.1.0
cryptography==36.0.2
cycler==0.10.0
cytoolz==0.11.0
dateparser==1.0.0
db-dtypes==1.0.4
decompyle3==3.9.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
distlib==0.3.2
docutils==0.16
dydx-v3-python==1.0.10
ecdsa==0.16.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.4
eth-brownie==1.12.4
eth-event==1.2.1
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
execnet==1.9.0
filelock==3.0.12
fixedint==0.1.6
flake8==3.8.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozendict==1.2
frozenlist==1.2.0
gevent==21.12.0
google-api-core==2.10.2
google-auth==2.6.2
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-logging==3.0.0
google-cloud-pubsub==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.50.0
grpcio-status==1.44.0
gunicorn==20.1.0
helpers==0.2.0
hexbytes==0.2.1
hyperlink==21.0.0
hypothesis==6.27.3
idna==3.3
imagesize==1.2.0
importlib-metadata==4.8.2
importlib-resources==5.6.0
incremental==21.3.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.7.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.7.0
itsdangerous==2.1.2
jedi==0.16.0
Jinja2==2.11.3
joblib==0.15.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-server==2.6.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.7
lxml==4.9.1
Mako==1.2.0
Markdown==3.1.1
MarkupSafe==2.0.1
marshmallow==3.5.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.7.0
mpl-finance==0.10.1
mplfinance==0.12.9b5
mpmath==1.0.0
msgpack==0.6.2
multiaddr==0.0.9
multidict==5.2.0
multitasking==0.0.11
munch==2.5.0
mypy-extensions==0.4.3
mythx-models==1.9.1
nbclassic==0.3.1
nbconvert==5.6.1
nbformat==5.0.4
netaddr==0.8.0
notebook==6.0.3
numpy==1.18.4
packaging==21.3
pandas==1.0.3
pandas-datareader==0.8.1
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.6.2
pathspec==0.9.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
polygon-api-client==0.2.11
prometheus-client==0.7.1
prompt-toolkit==3.0.8
property-cached==1.6.4
proto-plus==1.22.1
protobuf==4.21.8
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.6.0
py==1.11.0
py-solc-ast==1.2.8
py-solc-x==1.1.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycodestyle==2.6.0
pycoingecko==2.2.0
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.12.0
pyflakes==2.2.0
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
PyLD==2.0.1
pylint==2.5.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.20.0
python-http-client==3.3.7
pythx==1.6.1
pytz==2020.1
PyYAML==5.4.1
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
Quandl==3.5.0
redis==4.2.2
regex==2021.11.10
requests==2.28.1
requests-mock==1.6.0
requests-unixsocket==0.2.0
rlp==1.2.0
rsa==4.8
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.6
scikit-learn==0.23.0
scipy==1.4.1
semantic-version==2.8.5
Send2Trash==1.5.0
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
spark-parser==1.8.9
Sphinx==3.1.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.34
starkbank-ecdsa==2.0.3
statsmodels==0.11.1
sympy==1.6
ta==0.7.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tiingo==0.12.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tox==3.13.2
tqdm==4.53.0
traitlets==4.3.3
Twisted==22.2.0
txaio==22.2.1
typed-ast==1.5.1
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.6.0
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.21.0
webargs==5.5.3
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.1.0
websockets==9.1
Werkzeug==2.1.1
widgetsnbextension==3.5.1
wrapt==1.13.3
xdis==6.0.4
xlrd==1.2.0
yarl==1.7.2
yfinance==0.1.87
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
